feat(system): implement multi-level logger and fix UI crash
This commit introduces a comprehensive, multi-level logging system to replace the previous boolean `debug` flag. It also resolves a critical crash in the floating text edit panel.

**Key Changes:**

- **New Logging System:**
  - Implemented a numeric `logLevel` system (0-ERROR, 1-WARN, 2-INFO, 3-DEBUG) in `logger.js`.
  - Created a `formatError` utility to prevent `[object Object]` errors by properly serializing error stacks.
  - Replaced the `debug` flag in `config.js` with `logLevel`.
  - Updated `main.js` and `configManagerFactory.js` to initialize and use the new system.

- **New Commands:**
  - Created `/log <level>` to allow admins to set the log verbosity.
  - Created `/debug` as a quick toggle between INFO and DEBUG levels.
  - Renamed the old `debug.js` to `log.js`.

- **Bug Fix:**
  - Fixed a crash in `floatingTextEditPanel` by correcting the arguments passed to the `.slider()` method. The `defaultValue` is now passed as a raw number instead of an object.
  - Added clamping and validation for the slider's default value to prevent crashes from legacy data.

- **CI/CD:**
  - Updated the `.github/workflows/release.yml` to set `logLevel: 3` for beta releases.

+++ AddonExeBP/scripts/modules/commands/log.js
@@ -16,7 +16,7 @@ commandManager.register({
     permissionLevel: 1, // Admin and above
     allowConsole: true,
     parameters: [
-        { name: 'level', type: 'float', description: 'The log level to set (0-3).', optional: true }
+        { name: 'level', type: 'int', description: 'The log level to set (0-3).', optional: true }
     ],
     execute: (player, args) => {
         const { level } = args;
